Bug fixes + minor changes
For the content date section, can we change the input format as text? We’d like to use that section for platforms (Busrides vs. GCcampus)
For the reading time section, can we remove the “read” text since we have some products that are courses.
For the reading time section, as we have products that’s longer than 1 hour while some are just couple minutes, is it possible to change that to [x] hr [x] min? or probably we can change to text input and I’ll plug them in as text.
The description text in the night mode is in colour black, can we bring that to the colour that’s the same as the reading time section?